Moved shared driver setup to common usage
