+++ acceptance-tests-common/src/main/java/com/opencredo/test/SharedDriver.java
@@ -1,4 +1,4 @@
-package com.opencredo.test.ui.acceptance.test.config.webdriver;
+package com.opencredo.test;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -6,15 +6,24 @@ import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.remote.BrowserType;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.support.events.EventFiringWebDriver;
+import org.openqa.selenium.Proxy;
+import org.openqa.selenium.remote.CapabilityType;
 
 /**
  * Creates a browser session which can be used to interact with a web UI.
  * This class allows for a single session to be re-used across multiple
  * test cases for increased speed.
  */
+
 public class SharedDriver extends EventFiringWebDriver {
     private static WebDriver REAL_DRIVER;
     private static final Thread CLOSE_THREAD = new Thread(SharedDriver::quitGlobalInstance);
+
+    public SharedDriver(String browser, boolean isProxyEnabled, Proxy proxy) {
+        super(getRealDriver(browser, isProxyEnabled, proxy));
+        Runtime.getRuntime().addShutdownHook(CLOSE_THREAD);
+    }
+
 
     private static void quitGlobalInstance() {
         WebDriver driver = REAL_DRIVER;
@@ -24,8 +33,11 @@ public class SharedDriver extends EventFiringWebDriver {
         }
     }
 
-    private static WebDriver getRealDriver(String browser) {
+    private static WebDriver getRealDriver(String browser, boolean isProxyEnabled, Proxy proxy) {
         DesiredCapabilities capabilities = new DesiredCapabilities();
+        if (isProxyEnabled && proxy != null)
+            capabilities.setCapability(CapabilityType.PROXY, proxy);
+
         if (REAL_DRIVER == null) {
             switch (browser) {
                 case BrowserType.FIREFOX:
@@ -41,11 +53,6 @@ public class SharedDriver extends EventFiringWebDriver {
         return REAL_DRIVER;
     }
 
-    public SharedDriver(String browser) {
-        super(getRealDriver(browser));
-        Runtime.getRuntime().addShutdownHook(CLOSE_THREAD);
-    }
-
     @Override
     public void close() {
         if (Thread.currentThread() != CLOSE_THREAD) {
@@ -58,5 +65,3 @@ public class SharedDriver extends EventFiringWebDriver {
         }
     }
 }
-
-
