+++ security-acceptance-tests/src/test/resources/props-for-security-tests.properties
@@ -5,3 +5,4 @@ security.acceptance.test.selenium.wait.timeout.seconds=10
 security.acceptance.test.proxy.enabled=true
 security.acceptance.test.proxy.url=localhost
 security.acceptance.test.proxy.port=8888
+security.acceptance.test.browser=googlechrome
