+++ ui-acceptance-tests/src/test/java/com/opencredo/test/ui/acceptance/test/config/spring/UiTestConfig.java
@@ -1,9 +1,9 @@
 package com.opencredo.test.ui.acceptance.test.config.spring;
 
+import com.opencredo.test.SharedDriver;
 import com.opencredo.test.ui.acceptance.test.interaction.objects.GoogleSearchPageObject;
 import com.opencredo.test.ui.acceptance.test.config.TestWorld;
 import com.opencredo.test.ui.acceptance.test.config.UiTestProperties;
-import com.opencredo.test.ui.acceptance.test.config.webdriver.SharedDriver;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
@@ -19,7 +19,7 @@ public class UiTestConfig {
 
     @Bean
     public SharedDriver driver() {
-        return new SharedDriver(properties().getBrowser());
+        return new SharedDriver(properties().getBrowser(), false, null);
     }
 
     @Bean
