+++ ui-acceptance-tests/src/test/java/com/opencredo/test/ui/acceptance/test/config/webdriver/WaitConditions.java
@@ -23,12 +23,9 @@ public class WaitConditions {
 
     public static ExpectedCondition<Boolean> pageContainsText(final String text) {
         return new ExpectedCondition<Boolean>() {
-            private String currentPage = "";
-
             @Override
             public Boolean apply(WebDriver driver) {
-                currentPage = driver.getPageSource();
-                return currentPage.contains(text);
+                return  driver.getPageSource().contains(text);
             }
 
             @Override
